Refactored loadOffers action
