+++ project/src/types/json-value/json-value.ts
@@ -1,0 +1,3 @@
+type JSONValue = { [key: string]: number | boolean | string | string[] | JSONValue };
+
+export default JSONValue;
