+++ project/src/utils/index.ts
@@ -1,3 +1,4 @@
 export {default as createOffersCityMap} from './create-offers-city-map/create-offers-city-map';
 export {default as findCityOffers} from './find-city-offers/find-city-offers';
 export {default as sortOffers} from './sort-offers/sort-offers';
+export {default as parseOffer} from './parse-offer/parse-offer';
